update TotalsNumber to vitest
